Implement communication via shared mem, UI via tui
The new UI of the broker will not be in the browser anymore, but in the
terminal. This step helps to improve performance alot.
The UI is still missing important features and a nice style, but it does
its job for now.
Crates 'crossterm' + 'tui' will handle user input and rendering.

The communication via shared memory causes some unresolved panics, which
must be looked into.